+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/common/entities/message_types/OutrightLeagueSettlementUpdate.java
@@ -1,25 +1,26 @@
 package eu.lsports.trade360_java_sdk.common.entities.message_types;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import eu.lsports.trade360_java_sdk.common.entities.markets.MarketEvent;
 import eu.lsports.trade360_java_sdk.common.entities.outright_league.OutrightLeagueCompetition;
+import eu.lsports.trade360_java_sdk.common.entities.outright_league.OutrightLeagueMarketEvent;
 import jakarta.annotation.Nullable;
 
 /**
- * The {@code OutrightLeagueMarketUpdate} class represents an update to the markets of an outright league for trade360 customers.
+ * The {@code OutrightLeagueSettlementUpdate} class represents an update to the
+ * settlements of an outright league for trade360 customers.
  */
 public class OutrightLeagueSettlementUpdate {
 
     /**
-     * The entity key for the outright league market update.
+     * The entity key for the outright league settlement update.
      */
     @JsonIgnore
     public final static int entityKey = 43;
 
     /**
-     * The competition associated with the outright league market update.
+     * The competition associated with the outright league settlement update.
      * This can be {@code null} if not set.
      */
     @Nullable
-    public OutrightLeagueCompetition<MarketEvent> competition;
+    public OutrightLeagueCompetition<OutrightLeagueMarketEvent> competition;
 }