+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/common/entities/outright_league/OutrightLeagueMarketCompetition.java
@@ -1,6 +1,5 @@
 package eu.lsports.trade360_java_sdk.common.entities.outright_league;
 
-import eu.lsports.trade360_java_sdk.common.entities.markets.MarketEvent;
 import jakarta.annotation.Nullable;
 
 import java.time.Instant;
@@ -8,7 +7,7 @@ import java.time.Instant;
 /**
  * The {@code OutrightLeagueMarketCompetition} class represents a competition in an outright league market update (type 40).
  */
-public class OutrightLeagueMarketCompetition extends OutrightLeagueCompetition<MarketEvent> {
+public class OutrightLeagueMarketCompetition extends OutrightLeagueCompetition<OutrightLeagueMarketEvent> {
 
     /**
      * The start time of the next fixture in UTC.